+++ nix/consumer-cache-hash.nix
@@ -1,1 +1,1 @@
-"sha256-VsYoVgIiJac+AbGK6SjuPTar/zibVBUyqDkP0yLOXwQ="
+"sha256-EOhSmaRoTK9KOl1aipOh6pLWpZanH9rlqehhXoPflMQ="